+++ ghc-stack-profiler/src/GHC/Stack/Profiler/Internal/Stack/Decode.hs
@@ -1,6 +1,6 @@
 {-# LANGUAGE MagicHash #-}
 
-module GHC.Stack.Profiler.Stack.Decode (
+module GHC.Stack.Profiler.Internal.Stack.Decode (
   decodeStackWithIpProvId,
 ) where
 
@@ -18,7 +18,7 @@ import GHC.Stack.CloneStack (StackSnapshot (..))
 import GHC.Exts.Heap.InfoTable.Types
 
 import GHC.Stack.Profiler.Core (IpeId (..), SourceLocation (..), StackItem (..))
-import GHC.Stack.Profiler.Stack.Compat (lookupIpeIdForStackFrame)
+import GHC.Stack.Profiler.Internal.Stack.Compat (lookupIpeIdForStackFrame)
 
 decodeStackWithIpProvId :: StackSnapshot -> IO [StackItem]
 decodeStackWithIpProvId (StackSnapshot stack#) = do
